requirements.txt
annotated-types==0.6.0
anyio==4.3.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
fastapi==0.110.0
filelock==3.9.0
Flask==3.0.2
Flask-Cors==4.0.0
fsspec==2023.4.0
gunicorn==21.2.0
h11==0.14.0
httptools==0.6.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
packaging==23.2
pillow==10.2.0
pydantic==2.6.3
pydantic_core==2.16.3
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.1
starlette==0.36.3
sympy==1.12
torch
torchaudio
torchvision
tqdm==4.66.2
typing_extensions==4.8.0
urllib3==2.2.1
uvicorn==0.27.1
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
